// src/create/model/ApprovalContentsRdo.ts
export class ApprovalContentsRdo {
  //
  serviceType: string = '';
  cubeType: string = '';
  channel: string = '';
  college: string = '';
  openState: string = '';
  name: string = '';
  creatorName: string = '';
  startDate: number = 0;
  endDate: number = 0;
  limit: number = 0;
  offset: number = 0;
}